Move namespaces.
