+++ Core/DnsUpdater.Core.csproj
