+++ Core/Services/IDnsRecordUpdater.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net;
 using System.Threading.Tasks;
 
-namespace Core.Services
+namespace DnsUpdater.Core.Services
 {
     public interface IDnsRecordUpdater
     {
