+++ Host/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 ﻿{
   "profiles": {
-    "WindowsService": {
+    "DnsHost": {
       "commandName": "Project",
       "dotnetRunMessages": "true",
       "environmentVariables": {
