+++ IpResolvers.Ipify/DnsUpdater.IpResolvers.Ipify.csproj
@@ -9,7 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Core\Core.csproj" />
+    <ProjectReference Include="..\Core\DnsUpdater.Core.csproj" />
   </ItemGroup>
 
 </Project>
